+++ Kitodo/src/test/java/org/kitodo/production/forms/dataeditor/StructurePanelTest.java
@@ -14,9 +14,6 @@ package org.kitodo.production.forms.dataeditor;
 import java.lang.reflect.Field;
 import java.lang.reflect.Method;
 import java.net.URI;
-import java.util.Arrays;
-import java.util.LinkedList;
-import java.util.List;
 
 import org.junit.Assert;
 import org.junit.Test;
@@ -51,43 +48,5 @@ public class StructurePanelTest {
         buildStructureTreeRecursively.invoke(underTest, structure, result);
 
         Assert.assertTrue(((StructureTreeNode) result.getChildren().get(0).getData()).isLinked());
-    }
-
-    @Test
-    public void testDetermineIncludedStructuralElementPathToChildRecursive() throws Exception {
-        IncludedStructuralElement includedStructuralElement = new IncludedStructuralElement();
-        includedStructuralElement.setType("newspaperYear");
-
-        IncludedStructuralElement monthIncludedStructuralElement = new IncludedStructuralElement();
-        monthIncludedStructuralElement.setType("newspaperMonth");
-
-        IncludedStructuralElement wrongDayIncludedStructuralElement = new IncludedStructuralElement();
-        wrongDayIncludedStructuralElement.setType("newspaperDay");
-        wrongDayIncludedStructuralElement.setLabel("wrong");
-        LinkedMetsResource wrongLink = new LinkedMetsResource();
-        wrongLink.setUri(URI.create("database://?process.id=13"));
-        wrongDayIncludedStructuralElement.setLink(wrongLink);
-        monthIncludedStructuralElement.getChildren().add(wrongDayIncludedStructuralElement);
-
-        IncludedStructuralElement correctDayIncludedStructuralElement = new IncludedStructuralElement();
-        correctDayIncludedStructuralElement.setType("newspaperDay");
-        correctDayIncludedStructuralElement.setLabel("correct");
-        LinkedMetsResource correctLink = new LinkedMetsResource();
-        correctLink.setUri(URI.create("database://?process.id=42"));
-        correctDayIncludedStructuralElement.setLink(correctLink);
-        monthIncludedStructuralElement.getChildren().add(correctDayIncludedStructuralElement);
-
-        includedStructuralElement.getChildren().add(monthIncludedStructuralElement);
-        int number = 42;
-
-        Method determineIncludedStructuralElementPathToChildRecursive = StructurePanel.class.getDeclaredMethod(
-            "determineIncludedStructuralElementPathToChildRecursive", IncludedStructuralElement.class, int.class);
-        determineIncludedStructuralElementPathToChildRecursive.setAccessible(true);
-        @SuppressWarnings("unchecked")
-        List<IncludedStructuralElement> result = (List<IncludedStructuralElement>) determineIncludedStructuralElementPathToChildRecursive
-                .invoke(null, includedStructuralElement, number);
-
-        Assert.assertEquals(new LinkedList<>(Arrays.asList(includedStructuralElement, monthIncludedStructuralElement,
-            correctDayIncludedStructuralElement)), result);
     }
 }
